Manipulador-Planar/requirements.txt
pybullet
numpy
paho-mqtt
requests
